Improve feederservice/cargodest operation by subtracting the offset profit solely from the final vehicle, instead of the next vehicle, preventing intermediary services being hit by large profit penalties (previously: loss > than running costs occurred in some cases).
Change behaviour of packet unloader code when the station being unloaded at is full (32767 units with newcargos on).
Now old packets (or if no packets, a suitable amount of unrouted cargo), are discarded to make room. This prevents the station being deadlocked as every vehicle would wait for the station to desaturate before it could unload (which furthers saturation) and then load.