Make cargodest next hop target predictor now follow unconditional order skips. Also make some formatting and comment changes.
Conditional skips are currently still ignored by cargodest, hence caution should be used when using both features together.
(A solution to this, if both features are to be used together, until the above is successfully addressed, is to have the same next hop regardless of whether a conditional skip is taken, ie. by moving and duplicating the (unloadable) station prior to the conditional skip back through it, (this is probably only of use for branch conditions)).